libmusic: fix type casting in CQTWrapper::Process()
uint_16t is not enough to hold time-domain samples count, which can be way beyond uint_16t limit,
e.g. 2 seconds of 44.1 kHz track is 88200 samples while uint16_t limit is 65536.

Assuming a situation: 4096 window size, 4096 hop size, 88200 samples of TD data,
at 5th iteration we're at sample 20480. Then static_cast<uint16_t>(88200 - 20480) == 2184, and
hence we feed cqtt only first 2184 samples of window, completely skipping the rest 1912 samples.
IOW not all time domain input get processed which may impact spectrogram calculation.

+++ libmusic/src/cqt_wrapper.cpp
@@ -71,8 +71,8 @@ void CQTWrapper::Process(const td_t & td, uint32_t offset)
     CQBase::RealBlock output_block, output;
 
     if (hop_size_ > 1) {
-        for (uint32_t sample = offset; sample < td.size(); sample += hop_size_) {
-            uint32_t len = min(win_size_, static_cast<uint16_t>(td.size() - sample));
+        for (size_t sample = offset; sample < td.size(); sample += hop_size_) {
+            size_t len = min(static_cast<size_t>(win_size_), td.size() - sample);
             CQBase::RealSequence input(td.begin() + sample, td.begin() + sample + len);
             output_block = cq_spectrogram_->process(input);
